chore: clean up chatbot controller & create service
