+++ backend/app/api/v1/routes.py
@@ -5,7 +5,7 @@ from app.api.v1.endpoints.article_controller import router as article_router
 from app.api.v1.endpoints.article_matching_controller import (
     router as article_matching_router,
 )
-from app.api.v1.endpoints.chats_controller import router as chats_router
+from app.api.v1.endpoints.chatbot_controller import router as chatbot_router
 from app.api.v1.endpoints.crawler_controller import router as crawler_router
 from app.api.v1.endpoints.email_controller import router as email_router
 from app.api.v1.endpoints.keyword_controller import router as keyword_router
@@ -38,7 +38,7 @@ router_list = [
     keyword_router,
     crawler_router,
     report_router,
-    chats_router,
+    chatbot_router,
     organization_router,
 ]
