+++ backend/app/schemas/chatbot_schemas.py
@@ -1,0 +1,9 @@
+from pydantic import BaseModel
+
+
+class ChatRequest(BaseModel):
+    sender: str
+    subject: str
+    body: str
+    s3_key: str
+    bucket: str
